refactor(scraper): refactor scraping logic
